# Conflicts:
#	src/test/java/com/hhplus/concert_ticketing/business/facade/impl/ConcertInfoManagementFacadeImplIntegratedTest.java

+++ src/main/java/com/hhplus/concert_ticketing/presentation/dto/response/TokenResponse.java
@@ -9,11 +9,11 @@ import java.time.LocalDateTime;
 public class TokenResponse {
 
     String          token;
-    Long            queuePosition;
+    Integer         queuePosition;
     LocalDateTime   expired_at;
 
     @Builder
-    public TokenResponse(String token, Long queuePosition, LocalDateTime expired_at) {
+    public TokenResponse(String token, Integer queuePosition, LocalDateTime expired_at) {
         this.token = token;
         this.queuePosition = queuePosition;
         this.expired_at = expired_at;
